requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
gunicorn==21.2.0
SQLAlchemy==2.0.45
pydantic==2.12.5
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0